+++ tests/functional/adapter/dbt/test_empty.py
@@ -12,14 +12,21 @@ from dbt.tests.util import run_dbt
 
 model_sql_sqlserver = """
 select *
-from {{ ref('model_input') }}
+from {{ ref('model_input') }} as model_input_alias
 union all
 select *
-from {{ source('seed_sources', 'raw_source') }}
+from {{ source('seed_sources', 'raw_source') }} as raw_source_alias
 """
 
 model_inline_sql_sqlserver = """
-select * from {{ source('seed_sources', 'raw_source') }}
+select * from {{ source('seed_sources', 'raw_source') }} as raw_source_alias
+"""
+
+model_sql_user_alias_sqlserver = """
+select user_alias.id as id
+from {{ ref('model_input') }} as user_alias
+inner join {{ source('seed_sources', 'raw_source') }} as source_alias
+    on user_alias.id = source_alias.id
 """
 
 
@@ -43,6 +50,49 @@ class TestEmpty(BaseTestEmpty):
         self.assert_row_count(project, "model", 2)
 
         # run with empty - 0 expected rows in output
+        run_dbt(["run", "--empty"])
+        self.assert_row_count(project, "model", 0)
+
+
+class TestEmptyWithUserAlias(BaseTestEmpty):
+    @pytest.fixture(scope="class")
+    def models(self):
+        return {
+            "model_input.sql": model_input_sql,
+            "model.sql": model_sql_user_alias_sqlserver,
+            "sources.yml": schema_sources_yml,
+        }
+
+    def test_run_with_empty(self, project):
+        run_dbt(["seed"])
+
+        run_dbt(["run", "--empty"])
+        self.assert_row_count(project, "model", 0)
+
+
+@pytest.mark.xfail(
+    reason="Upstream dbt empty-mode alias handling needs to be contextual aware.",
+)
+class TestEmptyWithUserAliasAndNoAliasFlag(BaseTestEmpty):
+    @pytest.fixture(scope="class")
+    def models(self):
+        return {
+            "model_input.sql": model_input_sql,
+            "model.sql": model_sql_user_alias_sqlserver,
+            "sources.yml": schema_sources_yml,
+        }
+
+    @pytest.fixture(scope="class")
+    def project_config_update(self):
+        return {
+            "flags": {
+                "dbt_sqlserver_disable_empty_relation_aliases": False,
+            }
+        }
+
+    def test_run_with_empty(self, project):
+        run_dbt(["seed"])
+
         run_dbt(["run", "--empty"])
         self.assert_row_count(project, "model", 0)